Merged PR 3639: DAM Connector Not Working on Headless Deployments
The CM container needed the CSP transform updated - the path was incorrect in the transform file.

DAM_ExternalRedirectKey env variable added (comes with the OOTB connector image) but we will not use it since it will be confusing when having multiple Sitecore instances pointed to a single CH instance.

No AKS changes needed for this one.

Related work items: #6211